only admin updates or delete mail delivered commit

+++ app/controllers/imagegals_controller.rb
@@ -75,7 +75,10 @@ class ImagegalsController < ApplicationController
   def update
     respond_to do |format|
       if @imagegal.update(imagegal_params)
-        ImagegalMailer.update_imagegal(@imagegal).deliver_now
+        if current_user.admin?
+          ImagegalMailer.update_imagegal(@imagegal).deliver_now
+        end
+        
         format.html { redirect_to @imagegal, notice: 'Imagegal was successfully updated.' }
         format.json { render :show, status: :ok, location: @imagegal }
       else
@@ -121,7 +124,10 @@ class ImagegalsController < ApplicationController
     end
     
     respond_to do |format|
-      ImagegalMailer.destroy_imagegal(@imagegal).deliver_now
+      if current_user.admin?
+        ImagegalMailer.destroy_imagegal(@imagegal).deliver_now
+      end
+      
       format.html { redirect_to imagegals_url, notice: 'Imagegal was successfully destroyed.' }
       format.json { head :no_content }
     end
